Fix failure in -[BTEncryptionTest testRoundTrip] under arm64
This test was only passing on 32 bit by chance. Changing the source
string to `@"test dataa"` causes the test to fail on both 32 and 64 bit
architectures. The problem is that `+[NSString stringWithUTF8String:]`
expects a null-terminated, UTF-8 encoded C-string, but `encodedAesKey`
is not guaranteed to meet this requirement.

The solution is to use `-[NSString initWithBytes:length:encoding:]` and
specify the length of the data explicitly.