Improve error message or exception for improper manual Apple Pay integrations
If a developer does a manual integration with a dynamic framework
target, and neglects to set `BT_ENABLE_APPLE_PAY=1` in your framework
Build Settings, but do so in the app target, they'd receive a
run-time exception for "tokenizeApplePayPayment:completion: is not
defined".

This change should improve the error messaging in this integration
error case.