Disable Build Implicit Dependencies in Xcode
Build Implicit Dependencies setting in Xcode does a poor job of
determining the dependency graph of its targets, which causes a lot of
unnecessary building of unrelated targets and indeterminate build
failures.

Disabling this setting also makes it easier to detect project
misconfigurations, such as Target Dependencies for the Demo app not
being set up correctly. Some of these were addressed in this commit.