Use dispatch groups, early return
Semaphores will lock the calling thread, and in the current implementation,
it will happen every single time this method is invoked.
We're observing a small percentage of app freezes caused by this method.

The new implementation early returns if the config was already initialized.
Another way of implementing it is with dispatch_once(), but that wouldn't
allow clearing out the config and re-fetching (not implemented in this PR to
minimize changes).