Issue 485: Carthage Submodule Integration Dirty
## Issue:
https://github.com/braintree/braintree_ios/issues/485

## Problem:
When integrating braintree_ios sdk via Carthage (as submodules), consumer Applications repos are dirtied due to a SYMLink between the cloned carthage checkout directory, and the Applications root Carthage/Checkout directory.

This had not been an issue with our previous integration of Release 4.24.0, which didn't have the Cardinal dependency yet to reveal this issue.

## Why
When Carthage integrates dependencies, it inspects each dependency as that dependencies dependency, and eventually constructs a Cartfile.resolved include all dependencies in the graph.

As part of this resolution, all dependencies of dependencies are cloned to the top level which may then be shared with dependencies (if multiple dependencies reuse the same item). This is done via a SYMLink.

In our case, this exibited the behavior of a new file being created in the cloned braintree submodule:
<ProjectDir>/Carthage/Checkouts/braintree_ios/Carthage/Checkouts/CardinalMobile

This new SYMLINK left our project dirty. Given the SYMLink exists in the braintree_ios folder structure, the folder which needs this .gitignore is the braintree_ios repo.

By default, when creating new github repo's a .gitignore may be provided by Github, often with default items such as this as well as CocoaPods. In this case, this is mirroring what the CocoaPods Pods path is doing just above this change.

## Change:
- Include Carthage Directory .gitignore to play friendly with Carthage integrations.